# src/main/java/com/example/bankcards/dto/JwtResponse.java
package com.example.bankcards.dto;
import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class JwtResponse {
    private String token;
    private String type;
    private String username;
    private String role;
}